fix: use change set sharded store for management runs
Some reactivity issues and a bug when fetching history were a
consequence of using the FuncRuns pinia store, which is not namespaced
to the change set. This moves the management functionality over to its
own, change set specific store, which resolves these issues.

Fixes BUG-664
